Refactors proxy controller
