+++ pkg/controller/proxyconfig/config_test.go
@@ -1,0 +1,3 @@
+package proxyconfig
+
+// TODO: Create proxy controller unit tests.
