--- pkg/proxy/proxy_config_test.go
@@ -1,3 +1,0 @@
-package proxy
-
-// TODO: Create proxy controller unit tests.
